requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coverage==7.6.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-model-utils==4.5.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
setuptools==71.0.3
sqlparse==0.5.0
tzdata==2024.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2

